Implement tear_down_data_panel function
The Application.tear_down_data_panel() function now properly destroys the old sample data panel entries and clears the lists containing the objects.
It is designed in such a way that the destruction happens only AFTER confirming the selection of a new sample or after confirming the opening of a file from disk.