Remove an import of matplotlib.pyplot to fix Python 3.5 on Mac
Macs were experiencing an issue using Python 3.5 and Matplotlib 1.5.1 where
an Abort trap: 6 error was being reported after starting the application.

The following stackoverflow question,
http://stackoverflow.com/questions/32019556/matplotlib-crashing-tkinter-application,
suggests that the issue was caused by an importing matplotlib.pyplot without
changing to the appropriate backend. Since only the object-oriented MPL is used
in this application, the import of pyplot was simply removed.