Add examples, instructions, and remove todo
An examples folder was added containing files that demonstrate the capabilities
of Rp Plotter. There are minimal working examples of the current features.
The README.md file was updated with instructions for using the application and
with a description of the example files that are now included. The TODO section
at the top of RpPlotter.py was moved outside of the repository.